debug: above helper not removing citation and failing to parse http correctly.
