+++ examples/helper-macro-git/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "helper-macro-git"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+publish = { workspace = true }
+rust-version = { workspace = true }
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { workspace = true }
+quote = { workspace = true }
+proc-macro2 = { workspace = true }
+serde_json = { workspace = true }
+
+[lints]
+workspace = true
