+++ tests/helper-macro-git/Cargo.toml
@@ -9,15 +9,9 @@ homepage = { workspace = true }
 publish = { workspace = true }
 rust-version = { workspace = true }
 
-[lib]
-proc-macro = true
-
 [dependencies]
 cite = { workspace = true, features = ["warn"] }
-syn = { workspace = true }
-quote = { workspace = true }
-proc-macro2 = { workspace = true }
-serde_json = { workspace = true }
+helper-macro-git = { workspace = true }
 
 [dev-dependencies]
 trybuild = { workspace = true }
